# chore: remove unnecessary re-calculation in benchmark reporter