Recalculation to scale from seconds to microseconds isn't necessary.
The calculated stats can just be scaled. 

Fixes: #2064

* * *


To demonstrate no functional change:

### before

```
% ./test/benchmarks/scripts/run-benchmarks.sh ./test/benchmarks/004-transaction.js bench.out
transaction:
                             BENCH    CONTROL
ops/sec                      8,327     26,200
sample size                     54        158
sample arithmetic mean     120.088 μs  38.168 μs
sample variance             38.173      2.075
sample standard deviation    6.178 μs   1.440 μs
standard error of the mean   0.841 μs   0.115 μs
margin of error             ±1.648 μs  ±0.225 μs
relative margin of error    ±1.372 %   ±0.588 %
overhead                    81.920 μs
```

### after

```
% ./test/benchmarks/scripts/run-benchmarks.sh ./test/benchmarks/004-transaction.js bench.out
transaction:
                             BENCH    CONTROL
ops/sec                      9,478     28,354
sample size                     61        170
sample arithmetic mean     105.513 μs  35.268 μs
sample variance             16.699      1.458
sample standard deviation    4.086 μs   1.207 μs
standard error of the mean   0.523 μs   0.093 μs
margin of error             ±1.025 μs  ±0.182 μs
relative margin of error    ±0.972 %   ±0.515 %
overhead                    70.244 μs
```