@trentm 2021-04-30
https://apm-ci.elastic.co/job/apm-agent-nodejs/job/apm-agent-nodejs-mbp/view/change-requests/job/PR-2070/2/ to do a Benchmark run to sanity check I haven't broken anything.