@trentm 2021-04-30
The benchmark ran successfully: https://apm-ci.elastic.co/blue/organizations/jenkins/apm-agent-nodejs%2Fapm-agent-nodejs-mbp/detail/PR-2070/2/pipeline/145